// packages/react/src/components/shared/command-palette/command-palette.types.ts
import type { ReactNode } from "react";

interface OptionsConfig {
  label: string;
  icon?: ReactNode;
  shortcut?: string;
  value?: string;
  onSelect?: (value: string) => void
}

interface OptionsBaseConfig {
  heading: string;
  items: OptionsConfig[];
}

export interface CommandPaletteProps {
  open?: boolean;
  onOpenChange?: (open: boolean) => void;
  placeholder?: string;
  fallback?: string;
  options: OptionsBaseConfig[];
  className?: string;
}